requirements.txt
django
django-extensions
django_rest_framework
pydotplus